Refactored Program.cs
