+++ Reserveringssysteem/Reserveringen.cs
@@ -5,6 +5,6 @@ namespace Reserveringssysteem
 {
     public class Reserveringen
     {
-        public static Func<string> genereerReserveringscode = () => new string(Enumerable.Repeat("ABCDEFGHIJKLMNPQRSTUVWXYZ123456789", 4).Select(s => s[new Random().Next(s.Length)]).ToArray());
+        public static Func<string> GenereerReserveringscode = () => new string(Enumerable.Repeat("ABCDEFGHIJKLMNPQRSTUVWXYZ123456789", 4).Select(s => s[new Random().Next(s.Length)]).ToArray());
     }
 }
